refactor: move IssuerService to application
